+++ reflex/compiler/templates.py
@@ -651,209 +651,6 @@ def _render_hooks(
     return hooks_code
 
 
-def _custom_components_pyproject_toml_template(**kwargs) -> str:
-    """Template for custom components pyproject.toml.
-
-    Args:
-        **kwargs: Template context variables including package_name, module_name, reflex_version.
-
-    Returns:
-        Rendered pyproject.toml content as string.
-    """
-    package_name = kwargs.get("package_name", "")
-    module_name = kwargs.get("module_name", "")
-    reflex_version = kwargs.get("reflex_version", "")
-
-    return f"""[build-system]
-requires = ["setuptools", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "{package_name}"
-version = "0.0.1"
-description = "Reflex custom component {module_name}"
-readme = "README.md"
-license = {{ text = "Apache-2.0" }}
-requires-python = ">=3.10"
-authors = [{{ name = "", email = "YOUREMAIL@domain.com" }}]
-keywords = ["reflex","reflex-custom-components"]
-
-dependencies = ["reflex>={reflex_version}"]
-
-classifiers = ["Development Status :: 4 - Beta"]
-
-[project.urls]
-
-[project.optional-dependencies]
-dev = ["build", "twine"]
-
-[tool.setuptools.packages.find]
-where = ["custom_components"]
-"""
-
-
-def _custom_components_readme_template(**kwargs) -> str:
-    """Template for custom components README.
-
-    Args:
-        **kwargs: Template context variables including module_name, package_name.
-
-    Returns:
-        Rendered README.md content as string.
-    """
-    module_name = kwargs.get("module_name", "")
-    package_name = kwargs.get("package_name", "")
-
-    return f"""# {module_name}
-
-A Reflex custom component {module_name}.
-
-## Installation
-
-```bash
-pip install {package_name}
-```
-"""
-
-
-def _custom_components_source_template(**kwargs) -> str:
-    """Template for custom components source.
-
-    Args:
-        **kwargs: Template context variables including component_class_name, module_name.
-
-    Returns:
-        Rendered custom component source code as string.
-    """
-    component_class_name = kwargs.get("component_class_name", "")
-    module_name = kwargs.get("module_name", "")
-
-    return rf'''
-"""Reflex custom component {component_class_name}."""
-
-# For wrapping react guide, visit https://reflex.dev/docs/wrapping-react/overview/
-
-import reflex as rx
-
-# Some libraries you want to wrap may require dynamic imports.
-# This is because they they may not be compatible with Server-Side Rendering (SSR).
-# To handle this in Reflex, all you need to do is subclass `NoSSRComponent` instead.
-# For example:
-# from reflex.components.component import NoSSRComponent
-# class {component_class_name}(NoSSRComponent):
-#     pass
-
-
-class {component_class_name}(rx.Component):
-    """{component_class_name} component."""
-
-    # The React library to wrap.
-    library = "Fill-Me"
-
-    # The React component tag.
-    tag = "Fill-Me"
-
-    # If the tag is the default export from the module, you must set is_default = True.
-    # This is normally used when components don't have curly braces around them when importing.
-    # is_default = True
-
-    # If you are wrapping another components with the same tag as a component in your project
-    # you can use aliases to differentiate between them and avoid naming conflicts.
-    # alias = "Other{component_class_name}"
-
-    # The props of the React component.
-    # Note: when Reflex compiles the component to Javascript,
-    # `snake_case` property names are automatically formatted as `camelCase`.
-    # The prop names may be defined in `camelCase` as well.
-    # some_prop: rx.Var[str] = "some default value"
-    # some_other_prop: rx.Var[int] = 1
-
-    # By default Reflex will install the library you have specified in the library property.
-    # However, sometimes you may need to install other libraries to use a component.
-    # In this case you can use the lib_dependencies property to specify other libraries to install.
-    # lib_dependencies: list[str] = []
-
-    # Event triggers declaration if any.
-    # Below is equivalent to merging `{{ "on_change": lambda e: [e] }}`
-    # onto the default event triggers of parent/base Component.
-    # The function defined for the `on_change` trigger maps event for the javascript
-    # trigger to what will be passed to the backend event handler function.
-    # on_change: rx.EventHandler[lambda e: [e]]
-
-    # To add custom code to your component
-    # def _get_custom_code(self) -> str:
-    #     return "const customCode = 'customCode';"
-
-
-{module_name} = {component_class_name}.create
-'''
-
-
-def _custom_components_init_template(**kwargs) -> str:
-    """Template for custom components __init__.py.
-
-    Args:
-        **kwargs: Template context variables including module_name.
-
-    Returns:
-        Rendered __init__.py content as string.
-    """
-    module_name = kwargs.get("module_name", "")
-    return f"from .{module_name} import *"
-
-
-def _custom_components_demo_app_template(**kwargs) -> str:
-    """Template for custom components demo app.
-
-    Args:
-        **kwargs: Template context variables including custom_component_module_dir, module_name.
-
-    Returns:
-        Rendered demo app source code as string.
-    """
-    custom_component_module_dir = kwargs.get("custom_component_module_dir", "")
-    module_name = kwargs.get("module_name", "")
-
-    return rf'''
-"""Welcome to Reflex! This file showcases the custom component in a basic app."""
-
-from rxconfig import config
-
-import reflex as rx
-
-from {custom_component_module_dir} import {module_name}
-
-filename = f"{{config.app_name}}/{{config.app_name}}.py"
-
-
-class State(rx.State):
-    """The app state."""
-    pass
-
-def index() -> rx.Component:
-    return rx.center(
-        rx.theme_panel(),
-        rx.vstack(
-            rx.heading("Welcome to Reflex!", size="9"),
-            rx.text(
-                "Test your custom component by editing ",
-                rx.code(filename),
-                font_size="2em",
-            ),
-            {module_name}(),
-            align="center",
-            spacing="7",
-        ),
-        height="100vh",
-    )
-
-
-# Add state and page to the app.
-app = rx.App()
-app.add_page(index)
-'''
-
-
 # Template instances
 class TemplateFunction:
     """Wrapper for template functions to match Jinja Template interface."""
@@ -928,20 +725,3 @@ MACROS = TemplateFunction(
         kwargs.get("memo", None),
     )
 )
-
-# Code that generate the pyproject.toml file for custom components.
-CUSTOM_COMPONENTS_PYPROJECT_TOML = TemplateFunction(
-    _custom_components_pyproject_toml_template
-)
-
-# Code that generates the README file for custom components.
-CUSTOM_COMPONENTS_README = TemplateFunction(_custom_components_readme_template)
-
-# Code that generates the source file for custom components.
-CUSTOM_COMPONENTS_SOURCE = TemplateFunction(_custom_components_source_template)
-
-# Code that generates the init file for custom components.
-CUSTOM_COMPONENTS_INIT_FILE = TemplateFunction(_custom_components_init_template)
-
-# Code that generates the demo app main py file for testing custom components.
-CUSTOM_COMPONENTS_DEMO_APP = TemplateFunction(_custom_components_demo_app_template)
